+++ server/ngserver-java/drive-common/src/main/java/drive_protocol/response/ReadFileResponseOrBuilder.java
@@ -8,10 +8,10 @@ public interface ReadFileResponseOrBuilder extends
     com.google.protobuf.MessageOrBuilder {
 
   /**
-   * <code>int64 status = 1;</code>
+   * <code>int32 status = 1;</code>
    * @return The status.
    */
-  long getStatus();
+  int getStatus();
 
   /**
    * <code>bytes buffer = 2;</code>
