+++ server/ngserver-java/drive-common/src/main/java/drive_protocol/response/SetAllocationSizeResponseOrBuilder.java
@@ -8,8 +8,8 @@ public interface SetAllocationSizeResponseOrBuilder extends
     com.google.protobuf.MessageOrBuilder {
 
   /**
-   * <code>int64 status = 1;</code>
+   * <code>int32 status = 1;</code>
    * @return The status.
    */
-  long getStatus();
+  int getStatus();
 }
